Moved and renamed dialects, updated version to most recent release ones and added filter examples in the documentation
