+++ pom.xml
@@ -19,7 +19,6 @@
         </license>
     </licenses>
 
-
     <developers>
         <developer>
             <name>Codexio Team</name>
@@ -38,18 +37,18 @@
 
     <properties>
         <java.version>21</java.version>
-        <spring.web.version>5.3.18</spring.web.version>
+        <spring.web.version>6.1.6</spring.web.version>
         <spring.orm.version>6.1.6</spring.orm.version>
         <spring.data.jpa.version>3.2.5</spring.data.jpa.version>
         <spring.doc.version>2.5.0</spring.doc.version>
         <hibernate.core.version>6.4.1.Final</hibernate.core.version>
-        <jackson.databing.version>2.16.0</jackson.databing.version>
-        <jakarta.servlet.version>5.0.0</jakarta.servlet.version>
+        <jackson.databing.version>2.17.1</jackson.databing.version>
+        <jakarta.servlet.version>6.0.0</jakarta.servlet.version>
         <slf4j.version>2.0.13</slf4j.version>
-        <maven.compiler.version>3.8.1</maven.compiler.version>
+        <maven.compiler.version>3.13.0</maven.compiler.version>
         <maven.surefire.plugin.version>3.2.5</maven.surefire.plugin.version>
         <junit.version>5.10.1</junit.version>
-        <mockito.version>5.8.0</mockito.version>
+        <mockito.version>5.11.0</mockito.version>
 
         <nexus.plugin.version>1.6.13</nexus.plugin.version>
         <gpg.plugin.version>3.1.0</gpg.plugin.version>
@@ -191,8 +190,8 @@
                             <excludes>
                                 <exclude>**/payload/*</exclude>
                                 <exclude>**/FieldInfo.class</exclude>
-                                <exclude>**/ExtendedMysqlDialect.class</exclude>
-                                <exclude>**/ExtendedPostgresqlDialect.class</exclude>
+                                <exclude>**/RequeryEnhancedMySQLDialect.class</exclude>
+                                <exclude>**/RequeryEnhancedPostgreSQLDialect.class</exclude>
                             </excludes>
                         </configuration>
                         <executions>
