+++ src/main/java/bg/codexio/springframework/data/jpa/requery/dialect/RequeryEnhancedPostgreSQLDialect.java
@@ -1,14 +1,15 @@
-package bg.codexio.springframework.data.jpa.requery.resolver.function;
+package bg.codexio.springframework.data.jpa.requery.dialect;
 
+import bg.codexio.springframework.data.jpa.requery.resolver.function.CaseInsensitiveLikeSQLFunction;
 import org.hibernate.boot.model.FunctionContributions;
-import org.hibernate.dialect.MySQLDialect;
+import org.hibernate.dialect.PostgreSQLDialect;
 
 /**
- * Extends the standard {@link MySQLDialect} and registers a custom function
- * to support case-insensitive LIKE operations.
+ * Extends the standard {@link PostgreSQLDialect} and registers a custom
+ * function to support case-insensitive LIKE operations.
  */
-public class ExtendedMysqlDialect
-        extends MySQLDialect {
+public class RequeryEnhancedPostgreSQLDialect
+        extends PostgreSQLDialect {
     @Override
     public void initializeFunctionRegistry(FunctionContributions functionContributions) {
         super.initializeFunctionRegistry(functionContributions);
